+++ src/pybamm/meshes/meshes.py
@@ -232,11 +232,11 @@ class Mesh(dict):
             if (
                 len(domain) == 1
                 and not isinstance(
-                    submesh,
                     (
-                        pybamm.SubMesh0D,
-                        pybamm.ScikitSubMesh2D,
-                        pybamm.ScikitFemSubMesh3D,
+                        submesh,
+                        pybamm.SubMesh0D
+                        | pybamm.ScikitSubMesh2D
+                        | pybamm.ScikitFemSubMesh3D,
                     ),
                 )
             )